🚚 Simplify dbt tasks & utils structure
